Fixed a bug where when using R to go in reverse, it instantly kills you.
This was due to the frame buffer not being flushed since the last frame's erasure of the tail... making it seem like it's a black/free space but according to the buffer it's still your color.

This allows me to let people use R even when going fast or slow, now, too.

Still creates holes for connected players, sometimes.