Pathfinding for apples. Apples should be more fair now.
If the last person alive is in a tight spot, it will put the apple somewhere they can get it, unless they are in TOO tight of a spot, in which case it places it somewhere random.

Basically, if the box is only 20 x 20 big, it may not rez the apple inside it, but if its bigger, it will put it somewhere you an get it.