Fixed the bug plaguing me. The bug that caused the o3ds to hang often, but the n3ds very rarely, has been fixed.
The problem lied in the console being drawn in a separate frame from the top screen during the game_update(). The console and the top screen both get drawn in the same frame now, which has drastically increased the framerate of the game, and solved the problem of the 3ds hanging sometimes, seemingly at random.

Because this app uses system font, there is a buffer for the characters to be drawn. I hijacked this buffer for drawing snakes too. However, the array index only gets reset to zero when the bottom screen starts to be drawn on ITS frame, not the top screen too.

I replaced all calls of C3D_FrameBegin with myFrameBegin, which sets the index to zero, and calls C3D_FrameBegin for me.